% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_method.R
\name{CVRMSE}
\alias{CVRMSE}
\title{Return CVRMSE of the estimate}
\usage{
CVRMSE(y, y_hat, n_par)
}
\arguments{
\item{y}{the vector of the response variable}

\item{y_hat}{the vector of fitted values}

\item{n_par}{number of parameters in the model}
}
\description{
This function computes the cvrmse of the estimate
}
