% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process.R
\name{generate_lean_tex}
\alias{generate_lean_tex}
\title{Generate lean tex file}
\usage{
generate_lean_tex(plotType, region, topn, botn, category)
}
\arguments{
\item{topn}{top n records to include}

\item{botn}{bottom n records to include}

\item{category}{optional, filter for specific category}

\item{df}{required, a data frame with the image identifiers, and a column named "score"}

\item{isDesc}{whether the sort is descending}
}
\description{
This function generates the include image part of lean image tex file
}
\examples{
polynomial_deg_2(y, x)
}
\keyword{image}
\keyword{tex}
\keyword{to}
