% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_routine.R
\name{get_energy_weather_file}
\alias{get_energy_weather_file}
\title{Get energy and temperature data for building}
\usage{
get_energy_weather_file(id, latitude, longitude, lat_lon_df,
  debugFlag = FALSE)
}
\arguments{
\item{id}{required, Building ID (Building_Number), used in output filename}

\item{latitude}{numeric, latitude of the building}

\item{longitude}{numeric, longitude of the building}

\item{lat_lon_df}{if latitude and longitude are missing, user needs to input
a dataframe containing three columns: latitude, longitude, Building_Number (an id uniquely identifies the building)}

\item{debugFlag}{optional, flag of debugging, with temp file saved to db_build_temp_csv}
}
\description{
This function get energy and temperature of a building into a data frame
  eui_elec(electricity kBtu/sqft), eui_gas(gas kBtu/sqft)
}
\examples{
lean_analysis(lat_lon_df, radius=100, limit=5)
}
\keyword{lean}
