% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_method.R
\name{piecewise_linear}
\alias{piecewise_linear}
\title{Piecewise Linear Regression}
\usage{
piecewise_linear(y, x, h)
}
\arguments{
\item{y}{the vector of the response variable}

\item{x}{the vector of the 1-d covariate variable}

\item{h}{required, the h in the segmented::seg.control}
}
\description{
This function fits a piecewise linear regression
}
\examples{
piecewise_linear(y, x)
}
\keyword{linear}
\keyword{piecewise}
