% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_routine.R
\name{plot_lean_subset}
\alias{plot_lean_subset}
\title{Lean analysis for a subset of building}
\usage{
plot_lean_subset(region, buildingType, year, plotType, category)
}
\arguments{
\item{plotType}{"base", "elec", "gas"}

\item{category}{a vector of "A", "I", etc.}
}
\description{
This function tests the main lean analysis routine, with inputs from the db
}
\examples{
test_lean_analysis_db()
}
\keyword{lean}
\keyword{test}
