% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_method.R
\name{polynomial_deg_2}
\alias{polynomial_deg_2}
\title{Polynomial degree 2}
\usage{
polynomial_deg_2(y, x)
}
\arguments{
\item{y}{the vector of the response variable}

\item{x}{the vector of the 1-d covariate variable}
}
\description{
This function fits a polynomial of degree 2: y = b0 + b1x + b2x^2
}
\examples{
polynomial_deg_2(y, x)
}
\keyword{polynomial}
