% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_method.R
\name{test_fit}
\alias{test_fit}
\title{Testing fitting methods}
\usage{
test_fit()
}
\description{
This function tests various fitting methods implemented in this file
}
\examples{
polynomial_deg_2(y, x)
}
\keyword{fit}
\keyword{test}
