% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_routine.R
\name{test_lean_analysis_db}
\alias{test_lean_analysis_db}
\title{Lean analysis tester}
\usage{
test_lean_analysis_db()
}
\description{
This function tests the main lean analysis routine, with inputs from the db
}
\examples{
test_lean_analysis_db()
}
\keyword{lean}
\keyword{test}
