% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process.R
\name{img2tex}
\alias{img2tex}
\title{Image to tex}
\usage{
img2tex(df, prefix, suffix, isDesc, outfilename, topn)
}
\arguments{
\item{df}{required, a data frame with the image identifiers, and a column named "score"}

\item{isDesc}{whether the sort is descending}

\item{topn}{optional, if supplied, only write out the top n records}
}
\description{
This function compiles images in a data frame to a tex file, with sorting
}
\examples{
polynomial_deg_2(y, x)
}
\keyword{image}
\keyword{tex}
\keyword{to}
