% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_method.R
\name{plot_fit}
\alias{plot_fit}
\title{Plot the fit}
\usage{
plot_fit(yElec, yGas, x, resultElec, resultGas, plotType, id, methodName)
}
\arguments{
\item{x}{the vector of the 1-d covariate variable}

\item{resultElec}{the result returned by polynomial_deg_2 for electricity}

\item{resultGas}{the result returned by polynomial_deg_2 for gas}

\item{plotType}{the type of plots: "base_elec", "base_gas", "elec", "gas"}

\item{id}{optional, unique identifier of a building, default to "XXXXXXXX"}

\item{methodName}{optional, the name of the method}

\item{y}{the vector of the response variable}
}
\description{
This function plots the data and the fit line / curve
}
\examples{
plot_fit(y=df$`eui_elect`, x=df$`wt_temperatureFmonth`, output, color="red", methodName=NULL)
}
\keyword{polynomial}
