% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_routine.R
\name{plot_lean_subset}
\alias{plot_lean_subset}
\title{Lean analysis for a subset of building}
\usage{
plot_lean_subset(region, buildingType, buildingNumber, year, plotType, category,
  plotXLimit = NULL, plotYLimit = NULL, topn = NULL, botn = NULL,
  plotPoint = FALSE, elec_col, gas_col, debugFlag = FALSE, suffix = NULL)
}
\arguments{
\item{region}{optional, region number}

\item{buildingType}{optional, building type}

\item{year}{optional, restrict to data with fiscal year = year}

\item{plotType}{"base", "elec", "gas"}

\item{category}{optional, restrict to data with category in a vector of
categories, c(a vector of categories, e.g. "A", "I")}

\item{topn, }{optional, produce the topn highest score buildings' lean plot, commonly high score is bad usage}

\item{botn, }{optional, produce the botn lowest score buildings' lean plot}

\item{elec_col, }{required, column to plot the blue curve and yellow baseload}

\item{gas_col, }{required, column to plot the red curve and orange baseload}

\item{debugFlag, }{optional, whether to include additional debug columns in energy-weather csv file}

\item{suffix, }{optional, summary file suffixes}
}
\description{
This function produces the lean plot included in the regional report
}
\examples{
test_lean_analysis_db()
}
\keyword{lean}
\keyword{test}
