% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_routine.R
\name{plot_lean_subset}
\alias{plot_lean_subset}
\title{Lean analysis for a subset of building}
\usage{
plot_lean_subset(region, buildingType, buildingNumber, year, plotType, category,
  sourceEnergy = FALSE, plotXLimit = NULL, plotYLimit = NULL,
  topn = NULL, botn = NULL, plotPoint = FALSE)
}
\arguments{
\item{plotType}{"base", "elec", "gas"}

\item{category}{a vector of "A", "I", etc.}

\item{sourceEnergy, }{optional, whether to use source, instead of site energy}

\item{topn, }{optional, produce the topn score building's lean plot}

\item{botn, }{optional, produce the botn score building's lean plot}
}
\description{
This function produces the lean plot included in the regional report
}
\examples{
test_lean_analysis_db()
}
\keyword{lean}
\keyword{test}
