% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_routine.R
\name{stacked_fit_plot}
\alias{stacked_fit_plot}
\title{Lean analysis for a subset of building}
\usage{
stacked_fit_plot(region, buildingType, year, category, plotType, method,
  methodLabel)
}
\arguments{
\item{category}{a vector of "A", "I", etc.}

\item{plotType}{optional, "elec", "gas"}
}
\description{
This function tests the main lean analysis routine, with inputs from the db
}
\examples{
test_lean_analysis_db()
}
\keyword{lean}
\keyword{test}
